Edit qmail-local.c, fmt_xlong.c, and TARGETS
fmt_xlong.o was missing from TARGETS when its source was moved to its
own file.

fmt_xlong.c has been revised to move len, q, and c variables to one-per-
line in accordance with the changes made to genalloc in PR 128.

qmail-local.c has had the host[64] variable changed to a stralloc with a
readied size of 256. This is the same length reserved in djbdns-1.05 for
performing resolv.conf rewriting.